pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "loginal"
version = "0.1.0"
description = "A modular Python security log analysis framework"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Antigravity", email = "antigravity@gemini.google.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "rich",
    "python-evtx",
    "duckdb",
    "pandas",
    "pyyaml",
    "streamlit",
    "watchdog",
    "google-generativeai",
    "sentence-transformers",
    "faiss-cpu",
    "hdbscan",
    "umap-learn",
    "plotly",
    "scikit-learn",
]

[project.scripts]
loginal = "loginal.cli:main"
